telegram-bot/requirements.txt
python-telegram-bot==21.3
aiohttp==3.9.5
docker==7.1.0
